Improving Terminal Views
