+++ CodeEdit/Features/UtilityArea/TerminalUtility/TerminalListView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 import UniformTypeIdentifiers
 
 struct TerminalListView: View {
-    let group: UtilityAreaTerminalGroup
+    @State var group: UtilityAreaTerminalGroup
     @EnvironmentObject private var utilityAreaViewModel: UtilityAreaViewModel
     @FocusState.Binding var focusedTerminalID: UUID?
 
@@ -17,11 +17,39 @@ struct TerminalListView: View {
         VStack(spacing: 0) {
             ForEach(group.terminals, id: \.id) { terminal in
                 VStack(spacing: 0) {
-                    TerminalTabDragDropView(
+                    UtilityAreaTerminalRowView(
                         terminal: terminal,
-                        group: group,
                         focusedTerminalID: $focusedTerminalID
                     )
+                    .onDrag {
+                        utilityAreaViewModel.draggedTerminalID = terminal.id
+
+                        let dragInfo = TerminalDragInfo(terminalID: terminal.id)
+                        let provider = NSItemProvider()
+                        do {
+                            let data = try JSONEncoder().encode(dragInfo)
+                            provider.registerDataRepresentation(
+                                forTypeIdentifier: UTType.terminal.identifier,
+                                visibility: .all
+                            ) { completion in
+                                completion(data, nil)
+                                return nil
+                            }
+                        } catch {
+                            print("❌ Erro ao codificar dragInfo: \(error)")
+                        }
+                        return provider
+                    }
+                    .onDrop(
+                        of: [UTType.terminal.identifier],
+                        delegate: TerminalDropDelegate(
+                            groupID: group.id,
+                            viewModel: utilityAreaViewModel,
+                            destinationTerminalID: terminal.id
+                        )
+                    )
+                    .transition(.opacity.combined(with: .move(edge: .top)))
+                    .animation(.easeInOut(duration: 0.2), value: group.isCollapsed)
                 }
             }
         }
